initial state
managed to get the moving scripts going

+++ 2_copy_proj_after_pull.sh
@@ -1,19 +1,29 @@
 #!/bin/bash
 
-# Define the relative source and destination paths
-SOURCE_PATH="../project.pbxproj"
+# Define colors
+RED='\033[0;31m'
+YELLOW='\033[0;33m'
+GREEN='\033[0;32m'
+NC='\033[0m' # No Color
+
+# Define the source and destination paths
+SOURCE_PATH="backups/project.pbxproj"
 DESTINATION_PATH="../TouchChat.xcodeproj/project.pbxproj"
 
-# Check if the destination file exists and delete it
-if [ -f "$DESTINATION_PATH" ]; then
-    rm "$DESTINATION_PATH"
-    echo "Deleted existing file at $DESTINATION_PATH"
-else
-    echo "No existing file found at $DESTINATION_PATH"
+# Check if the source file exists
+if [ ! -f "$SOURCE_PATH" ]; then
+    echo -e "${RED}Source file $SOURCE_PATH does not exist. Halting operation.${NC}"
+    exit 1
 fi
 
-# Copy the file from the source to the destination
+# Check if the destination file exists and delete it if it does
+if [ -f "$DESTINATION_PATH" ]; then
+    rm "$DESTINATION_PATH"
+    echo -e "${YELLOW}Existing destination file $DESTINATION_PATH deleted.${NC}"
+fi
+
+# Copy the file from source to destination
 cp "$SOURCE_PATH" "$DESTINATION_PATH"
 
 # Print a message indicating that the copy was successful
-echo "File copied successfully from $SOURCE_PATH to $DESTINATION_PATH"
+echo -e "${GREEN}File copied successfully from $SOURCE_PATH to $DESTINATION_PATH${NC}"
